requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
flasgger==0.9.7.1
flask-cors==4.0.0
azure-cosmos==4.6.0
azure-communication-email==1.0.0
python-dotenv==1.0.1
requests==2.31.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
gunicorn==21.2.0
PyJWT==2.8.0
APScheduler==3.10.4
#weasyprint==65.1
fastapi==0.111.0
sqlalchemy==2.0.30
reportlab>=3.6.0
Flask-Limiter==3.5.0
faker>=20.0.0
pytest>=7.0.0
pytest-cov>=4.0.0